backend/requirements.txt
asgiref==3.5.0
black==22.3.0
click==8.1.2
colorama==0.4.4
Django==4.0.4
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
mypy-extensions==0.4.3
mysqlclient==2.1.0
pathspec==0.9.0
platformdirs==2.5.2
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
tzdata==2022.1
